\name{massRanges}
\alias{massRanges}
\title{Stores the m/z range}
\description{It stores the mass/charge ratio range as vector for later use}
\usage{massRanges(x,y)}
\arguments{
  \item{x}{- m/z 1}
  \item{y}{- m/z 2}
}
\author{Singh, Sukhdeep \email{sukhdeepsingh.bio@gmail.com}}
\section{Methods}{
\code{mass1=massRanges(200,300)}
}
\note{No notes yet}
\seealso{
	\code{\link{peakReader}}
	\code{\link{scanRanges}}
}
