\name{scanRanges}
\alias{scanRanges}
\title{Stores the scan range}
\description{It stores the scan range as a vector for later use}
\usage{scanRanges(x,y)}
\arguments{
  \item{x}{- range1}
  \item{y}{- range2}
}
\section{Methods}{\code{scan1=scanRanges(200,300)}}
\note{No notes yet}
\author{Singh, Sukhdeep \email{sukhdeepsingh.bio@gmail.com}}
\seealso{
	\code{\link{massRanges}}
	\code{\link{peakReader}}
	\code{\link{readData}}
	\code{\link{readFiles}}
}
